"Move api calls from app into seperate file"
